+++ src/main/java/kr/ync/service/FreeBoardService.java
@@ -2,7 +2,6 @@ package kr.ync.service;
 
 import java.util.List;
 
-import kr.ync.domain.BoardVO;
 import kr.ync.domain.Criteria;
 import kr.ync.domain.FreeBoardVO;
 
@@ -12,13 +11,13 @@ public interface FreeBoardService {
 	public void f_register(FreeBoardVO board);
 	
 	// 글 상세보기
-	public FreeBoardVO f_get(Long board_idx);
+	public FreeBoardVO f_get(int board_idx);
 	
 	// 글 수정
 	public boolean modify(FreeBoardVO board);
 	
 	// 글 삭제
-	public boolean remove(Long bno);
+	public boolean remove(int bno);
 	
 	// 전체 글 목록
 	public List<FreeBoardVO> getList();
